requirements.txt
streamlit
pillow
numpy
requests